+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.1.1] 29.12.2023
+- Added temperature parameter
+- Added support for GPT4-Turbo (gpt-4-1106-preview)
+- GPT3 now points to gpt-3.5-turbo-1106 (Updated GPT 3.5 Turbo)
+- NetSuiteChatGPTChat_UE cleanup
+
 ## [1.1.0] 28.12.2023
 - Added support for remembering previous messages (entire conversation is sent back to the API each time)
 - Conversations are saved in JSON format in the script folder